+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/betterOpPlayerNoCheat/ItemCommandMixin.java
@@ -42,7 +42,8 @@ public class ItemCommandMixin {
             //#else
             //$$ method = "method_13545",
             //#endif
-            remap = false, at = @At("RETURN"), cancellable = true)
+            remap = false, at = @At("RETURN"), cancellable = true
+    )
     private static void checkIfAllowCheating_itemCommand(ServerCommandSource source, CallbackInfoReturnable<Boolean> cir) {
         if (cir.getReturnValue() && CarpetAyakaSettings.betterOpPlayerNoCheat && !TISHelper.canCheat(source)) { // DO NOT change the order of the conditions
             cir.setReturnValue(false);