+++ changelog.md
@@ -1,6 +1,7 @@
 ## General
 
 - Added carpet rule fakePlayerForceOffline
+- Added carpet rule authProxyType, authProxyHost and authProxyPort
 - The despawn age of nether stars dropped by withers now becomes twice of itemDiscardAge
 - Added support for Minecraft 1.21.5 and 1.21.6
 - Fabric 0.16.0+ is now required